tests/test_console.py
#!/usr/bin/python3
"""Test Console module"""
import unittest
import pycodestyle
import inspect
import console


class Test_console(unittest.TestCase):
    """
    Test class for console
    All test cases writen in method like tests
    """

    @classmethod
    def setUpClass(cls) -> None:
        """Setup model instance to use in the test cases"""
        cls.fun_names = [name for name, _ in
                         inspect.getmembers(console.HBNBCommand,
                                            inspect.ismethod)]

    def test_documentation(self) -> None:
        """Test if module, class and methods documentations exist"""
        self.assertGreater(len(console.HBNBCommand.__doc__), 0)
        self.assertGreater(len(console.__doc__), 0)
        for func in self.fun_names:
            with self.subTest(func):
                self.assertGreater(len(func.__doc__), 0,
                                   f"Missing documentation of {func} method")

    def test_pep8_complaince(self) -> None:
        """Check if module is complaint when you run pycodestyle on it"""
        style_checker = pycodestyle.StyleGuide()
        result = style_checker.check_files(['console.py'])
        self.assertEqual(result.total_errors, 0, "PEP 8 violations found")


if __name__ == "__main__":
    unittest.main()
